menambahkan pembatasan pada php logic

+++ logic/delete_dosen.php
@@ -1,11 +1,16 @@
 <?php
 require_once("koneksi.php");
-$stmt = $pdo_conn->prepare("DELETE FROM dosen WHERE nip_dosen = :nip");
-$stmt->execute(array(
-    ':nip' => $_GET['nip']
-));
-$stmt = $pdo_conn->prepare("DELETE FROM akun WHERE username = :nip");
-$stmt->execute(array(
-    ':nip' => $_GET['nip']
-));
-header('location: ../list_dosen.php');
+session_start();
+if ($_SESSION['level_user'] != "Admin")
+    header('location: ../index.php');
+else {
+    $stmt = $pdo_conn->prepare("DELETE FROM dosen WHERE nip_dosen = :nip");
+    $stmt->execute(array(
+        ':nip' => $_GET['nip']
+    ));
+    $stmt = $pdo_conn->prepare("DELETE FROM akun WHERE username = :nip");
+    $stmt->execute(array(
+        ':nip' => $_GET['nip']
+    ));
+    header('location: ../list_dosen.php');
+}
